requirements.txt
selenium
webdriver-manager
openpyxl
pytest
allure-pytest

